Fix dropping a record on top of the next record
Dropping record on top of the following record in a list should be a no-op,
since dropping on a list item is the same as dropping before that item.
Now, we avoid doing it to avoid a QT crash.

Thanks to Kevin Otte <n8vnr@nivex.net> for discovering and reporting the crash.