Don't delete digital contact: All Call
I had used data from this page,
http://www.iz2uuf.net/wp/index.php/2016/06/04/tytera-dm380-codeplug-binary-format/,
which indicates a digital contact record beginning with 0xffffff is
deleted. Unfortunately, that is incorrect. An contact with a type of
"All Call" has a Call ID of 0xffffff (16777215). Such records were
incorrectly considered deleted.

We no longer consider those records as deleted.