Revert "Provide a standalone exe file instead of installer"
This reverts commit 15d67cf1b4864a443f7b03ad7ce3c10ee91db6a2.

It turns out that Windows is very picky about downloaded exe files
and presents a scare popup an indeterminate number of times for
such files.  We'll return to using an installer so that we get the
scare popup only once for the installer. It's also a bit nicer for
the typical windows user, I think.