Avoid checking sibling fields on creating records
This error occurred while importing JSON files.

Since we do not know the order of fields we'll be given when creating
a record, avoid checking the value of sibling fields that we may not
have yet.  After the record is created, check the validity of fields.

This commit fixes issue #23.